+++ Enums.cs
@@ -2,6 +2,7 @@
 {
     public static class Enums
     {
+        public enum eEmoteLoopingOptions { None, NetworkedLooping, LocalLoopingOnly }
         public enum eRopeVineChainOptions { Vanilla, AllowClimbing, AutoDismount }
         public enum eHotSunCookingOptions { None, OnlyFood, AllCookables }
     }